requirements.txt
Flask==3.1.2
python-dotenv==1.2.1
requests==2.32.5
gunicorn==21.2.0
